+++ Assets/CommonCoreGame/RPGGame/ObjectActions/ChangeSceneSpecial.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using CommonCore.ObjectActions;
 using CommonCore.World;
+using System;
 
 namespace CommonCore.RpgGame.ObjectActions
 {
@@ -14,6 +15,8 @@ namespace CommonCore.RpgGame.ObjectActions
     /// <para>Probably should not be triggerable by things other than player because it always affects player</para>
     /// <para>Adapted from DoorInteractableComponent</para>
     /// </remarks>
+    //added by mistake for Mother Earth
+    [Obsolete]
     public class ChangeSceneSpecial : ActionSpecial
     {
         public string Scene;
@@ -21,6 +24,8 @@ namespace CommonCore.RpgGame.ObjectActions
         public bool UsePosition = false;
         public Vector3 Position;
         public Vector3 Rotation;
+
+        public string TransferEffect;
 
         private bool Locked = false;
 
@@ -33,6 +38,12 @@ namespace CommonCore.RpgGame.ObjectActions
                     spawnPointHack = null;
                 else if (string.IsNullOrEmpty(SpawnPoint))
                     spawnPointHack = string.Empty;
+
+                if (!string.IsNullOrEmpty(TransferEffect))
+                {
+                    var targetTransform = data.Activator?.transform ?? transform;
+                    WorldUtils.SpawnEffect(TransferEffect, targetTransform.position, targetTransform.rotation, null, false);
+                }
 
                 WorldUtils.ChangeScene(Scene, spawnPointHack, Position, Rotation);
             }